informed-fe/pyproject.toml
[project]
name = "informed-fe"
version = "0.1.0"
description = "Food Ingredient Health Analyzer - Frontend UI"
requires-python = ">=3.11"
dependencies = [
    "streamlit>=1.36.0",
    "requests>=2.32.0",
    "pydantic>=2.11.0",
    "pydantic-settings>=2.10.0",
    "Pillow>=10.4.0",
    "prometheus-client>=0.20.0",  # Prometheus metrics
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["informed_fe*"]
